// src/model/item.ts
/**
 * API BHB Customer Backend
 * API BHB Customer Backend
 *
 * OpenAPI spec version: 1.0.0
 *
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
import { SubItem } from './subItem';

/**
 * Item information
 */
export interface Item {
  /**
   * Item id
   */
  id?: string;
  /**
   * Item name
   */
  name?: string;
  /**
   * Item plu
   */
  plu?: string;
  /**
   * Item price in cents without decimals
   */
  price?: number;
  /**
   * Item quantity
   */
  quantity?: number;
  subItems?: Array<SubItem>;
}